injector: fix array initializer
The struct insn_t's .bytes member is an array of length
MAX_INSN_LENGTH, which is #define'd to 15 in line 179 of injector.c.

The initialization of total_range in line 320 (which is of type
range_t, starring two members (.start and .end) of type insn_t,
however, uses 16 bytes for the initialization of the .bytes arrays.

Fix it by removing one element from the initializations.